% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp-parse-doc.R
\name{dp_parse_doc}
\alias{dp_parse_doc}
\title{Parse a document with a predefined parser}
\usage{
dp_parse_doc(file, parser, remote_id = NULL,
  docparser_api_key = Sys.getenv("DOCPARSER_API_KEY"))
}
\arguments{
\item{file}{path to a document (this will be \code{\link[=path.expanded]{path.expanded()}})}

\item{parser}{a parser id (which can be found via \code{\link[=dp_parsers]{dp_parsers()}})}

\item{remote_id}{if not \code{NULL} then an id which will be used to tag the file
on the \code{docparser} servers (otherwise one will be generated)}

\item{docparser_api_key}{if not placed in the \code{DOCPARSER_API_KEY} environment
variable, place it here. Otherwise, API requests fail.}
}
\value{
list of available parsers
}
\description{
Parse a document with a predefined parser
}
\examples{
\dontrun{
dp_parse_doc()
}
}
