% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp-get-parsers.R
\name{dp_parsers}
\alias{dp_parsers}
\title{Retrieve avilable parsers}
\usage{
dp_parsers(docparser_api_key = Sys.getenv("DOCPARSER_API_KEY"))
}
\arguments{
\item{docparser_api_key}{if not placed in the \code{DOCPARSER_API_KEY} environment
variable, place it here. Otherwise, API requests fail.}
}
\value{
data frame of available parsers
}
\description{
Retrieve avilable parsers
}
\examples{
\dontrun{
dp_parsers()
}
}
