% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp-ping.R
\name{dp_ping}
\alias{dp_ping}
\title{"Ping" the docparser API}
\usage{
dp_ping(docparser_api_key = Sys.getenv("DOCPARSER_API_KEY"))
}
\arguments{
\item{docparser_api_key}{if not placed in the \code{DOCPARSER_API_KEY} environment
variable, place it here. Otherwise, API requests fail.}
}
\value{
\code{TRUE} (invisibly) if successful
}
\description{
"Ping" the docparser API
}
\examples{
\dontrun{
dp_ping()
}
}
