% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp-get-results.R
\name{dp_results}
\alias{dp_results}
\title{Retrieve results from parsed documents}
\usage{
dp_results(doc_id, parser, format = c("object", "flat"),
  include_children = TRUE,
  docparser_api_key = Sys.getenv("DOCPARSER_API_KEY"))
}
\arguments{
\item{doc_id}{a document ID from the web interface or as returned by
\code{\link[=dp_parse_doc]{dp_parse_doc()}}}

\item{parser}{a parser id (which can be found via \code{\link[=dp_parsers]{dp_parsers()}})}

\item{format}{one of "\code{object}" or "\code{flat}". By default, parsed document
data is returned as nested JSON objects. Setting this parameter to
"\code{flat}" will return a simplified version of the parsed data which does
not contain flat key/value pairs instead of nested objects. Default
is "\code{object}".}

\item{include_children}{If child documents were created during preprocessing
(e.g. when splitting documents), setting this parameter to \code{TRUE}
(the default)
ensures that the parsed data of all child documents is returned.}

\item{docparser_api_key}{if not placed in the \code{DOCPARSER_API_KEY} environment
variable, place it here. Otherwise, API requests fail.}
}
\value{
list
}
\description{
Retrieve results from parsed documents
}
\examples{
\dontrun{
dp_results("484c4292ab3b324b7f2b60f3f2d3149f", "jcpfgqiexwky")
}
}
