% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docparser-package.R
\docType{package}
\name{docparser}
\alias{docparser}
\alias{docparser-package}
\title{...}
\description{
\itemize{
\item URL: \url{https://gitlab.com/hrbrmstr/docparser}
\item BugReports: \url{https://gitlab.com/hrbrmstr/docparser/issues}
}
}
\author{
Bob Rudis (bob@rud.is)
}
